update overdetermined/mixed signals test
