+++ src/dimensional_data.jl
@@ -13,3 +13,15 @@ function show(io::IO, mime::MIME{Symbol("text/plain")}, x::DimArray{Quantity{Flo
 end
 
 standard_error(P::AbstractDimArray{T,2}) where T <: Number = DimArray(.√diag(x.P),first(dims(x.P)))
+
+"""
+     function left divide
+
+     Left divide of Multipliable Matrix.
+     Reverse mapping from unitdomain to range.
+     Is `exact` if input is exact.
+"""
+function Base.:\(A::AbstractDimMatrix,b::AbstractDimVector)
+    DimensionalData.comparedims(first(dims(A)), first(dims(b)); val=true)
+    return rebuild(A,parent(A)\parent(b),(last(dims(A)),)) 
+end
